# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: URL Shortener API
  description: |
    A comprehensive API for URL shortening with analytics, user management, subscriptions, and payment processing.

    ## Authentication
    Most endpoints require authentication via JWT Bearer token or HTTP-only cookies.
    - **Access Token**: Short-lived token for API authentication
    - **Refresh Token**: Long-lived token for obtaining new access tokens

    ## Rate Limiting
    API requests may be rate limited based on your subscription plan.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:5080/api/v1
    description: Development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /plans:
    get:
      tags:
        - Plans
      summary: Get all plans
      description: Get all available subscription plans.
      operationId: getAllPlans
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: type
          in: query
          description: Filter by plan type
          schema:
            type: string
            enum: [free, basic, pro, enterprise]
        - name: isActive
          in: query
          description: Filter by active status
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: Plans retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanListResponse"
    post:
      tags:
        - Plans
      summary: Create plan (Admin)
      description: Create a new subscription plan. Admin only.
      operationId: createPlan
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanRequest"
      responses:
        "201":
          description: Plan created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /plans/slug/{slug}:
    get:
      tags:
        - Plans
      summary: Get plan by slug
      description: Get a plan by its URL-friendly slug.
      operationId: getPlanBySlug
      parameters:
        - name: slug
          in: path
          required: true
          description: Plan slug
          schema:
            type: string
      responses:
        "200":
          description: Plan retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /plans/admin:
    get:
      tags:
        - Plans
      summary: Get all plans (Admin)
      description: Get all plans including inactive ones. Admin only.
      operationId: getAllPlansAdmin
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: type
          in: query
          description: Filter by plan type
          schema:
            type: string
            enum: [free, basic, pro, enterprise]
        - name: isActive
          in: query
          description: Filter by active status
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: Plans retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /plans/seed:
    post:
      tags:
        - Plans
      summary: Seed plans (Admin)
      description: Seed default subscription plans. Admin only.
      operationId: seedPlans
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Plans seeded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /plans/{id}:
    get:
      tags:
        - Plans
      summary: Get plan by ID (Admin)
      description: Get a plan by its ID. Admin only.
      operationId: getPlanById
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/PlanId"
      responses:
        "200":
          description: Plan retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Plans
      summary: Update plan (Admin)
      description: Update a subscription plan. Admin only.
      operationId: updatePlan
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/PlanId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanRequest"
      responses:
        "200":
          description: Plan updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Plans
      summary: Delete plan (Admin)
      description: Delete a subscription plan. Admin only.
      operationId: deletePlan
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/PlanId"
      responses:
        "200":
          description: Plan deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ==================== SUBSCRIPTION ROUTES ====================
  /subscriptions/me:
    get:
      tags:
        - Subscriptions
      summary: Get my subscription
      description: Get the current user's active subscription.
      operationId: getMySubscription
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Subscription retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /subscriptions/me/history:
    get:
      tags:
        - Subscriptions
      summary: Get subscription history
      description: Get the current user's subscription history.
      operationId: getMySubscriptionHistory
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Subscription history retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionHistoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /subscriptions:
    post:
      tags:
        - Subscriptions
      summary: Subscribe to a plan
      description: Create a new subscription for the current user.
      operationId: subscribe
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionRequest"
      responses:
        "201":
          description: Subscription created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /subscriptions/cancel:
    post:
      tags:
        - Subscriptions
      summary: Cancel subscription
      description: Cancel the current user's subscription.
      operationId: cancelSubscription
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Subscription cancelled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No active subscription found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /subscriptions/change-plan:
    post:
      tags:
        - Subscriptions
      summary: Change subscription plan
      description: Change the current user's subscription to a different plan.
      operationId: changePlan
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionRequest"
      responses:
        "200":
          description: Plan changed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /subscriptions/admin:
    get:
      tags:
        - Subscriptions
      summary: Get all subscriptions (Admin)
      description: Get all subscriptions in the system. Admin only.
      operationId: getAllSubscriptions
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: status
          in: query
          description: Filter by subscription status
          schema:
            type: string
            enum: [active, cancelled, expired, pending, past_due]
        - name: userId
          in: query
          description: Filter by user ID
          schema:
            type: string
        - name: planId
          in: query
          description: Filter by plan ID
          schema:
            type: string
      responses:
        "200":
          description: Subscriptions retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /subscriptions/{id}:
    get:
      tags:
        - Subscriptions
      summary: Get subscription by ID (Admin)
      description: Get a subscription by its ID. Admin only.
      operationId: getSubscriptionById
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/SubscriptionId"
      responses:
        "200":
          description: Subscription retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Subscriptions
      summary: Update subscription (Admin)
      description: Update a subscription. Admin only.
      operationId: updateSubscription
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/SubscriptionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionRequest"
      responses:
        "200":
          description: Subscription updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /subscriptions/expire-check:
    post:
      tags:
        - Subscriptions
      summary: Check expired subscriptions (Admin)
      description: Run a check for expired subscriptions and update their status. Admin only.
      operationId: expireSubscriptions
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Expiration check completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ==================== PAYMENT ROUTES ====================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: HTTP-only cookie containing JWT access token

  # ==================== PARAMETERS ====================
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    PlanId:
      name: id
      in: path
      required: true
      description: Plan ID
      schema:
        type: string
    SubscriptionId:
      name: id
      in: path
      required: true
      description: Subscription ID
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            message: Authentication required
            statusCode: 401
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            message: Access denied. Admin privileges required.
            statusCode: 403
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            message: Resource not found
            statusCode: 404
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"

  # ==================== SCHEMAS ====================
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An error occurred
        statusCode:
          type: integer
          example: 400
      required:
        - success
        - message
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Invalid email format
      required:
        - success
        - message
        - errors
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
        hasNextPage:
          type: boolean
          example: true
        hasPrevPage:
          type: boolean
          example: false

    # Auth Schemas
...
    PlanFeatures:
      type: object
      properties:
        urlLimit:
          type: integer
          description: "-1 for unlimited"
          example: 100
        customAliasAllowed:
          type: boolean
          example: true
        analyticsEnabled:
          type: boolean
          example: true
        qrCodeEnabled:
          type: boolean
          example: true
        apiAccessEnabled:
          type: boolean
          example: false
        maxDevices:
          type: integer
          example: 3
        supportLevel:
          type: string
          enum: [basic, priority, dedicated]
          example: basic
    FeatureItem:
      type: object
      properties:
        id:
          type: string
          example: feat-1
        text:
          type: string
          example: Up to 100 short URLs
        order:
          type: integer
          example: 1
    Plan:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        id:
          type: string
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          example: Pro Plan
        slug:
          type: string
          example: pro-plan
        type:
          type: string
          enum: [free, basic, pro, enterprise]
          example: pro
        description:
          type: string
          example: Perfect for growing businesses
        features:
          $ref: "#/components/schemas/PlanFeatures"
        featuresList:
          type: array
          items:
            $ref: "#/components/schemas/FeatureItem"
        price:
          oneOf:
            - type: number
              example: 29.99
            - type: object
              properties:
                monthly:
                  type: number
                  example: 29.99
                yearly:
                  type: number
                  example: 299.99
        maxLinks:
          type: integer
          example: 500
        maxClicks:
          type: integer
          example: 50000
        interval:
          type: string
          enum: [month, year]
          example: month
        isActive:
          type: boolean
          example: true
        isDefault:
          type: boolean
          example: false
        sortOrder:
          type: integer
          example: 2
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    CreatePlanRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
          example: Pro Plan
        slug:
          type: string
          minLength: 2
          maxLength: 50
          pattern: "^[a-z0-9-]+$"
          example: pro-plan
        type:
          type: string
          enum: [free, basic, pro, enterprise]
          example: pro
        description:
          type: string
          example: Perfect for growing businesses
        features:
          $ref: "#/components/schemas/PlanFeatures"
        featuresList:
          type: array
          items:
            $ref: "#/components/schemas/FeatureItem"
        price:
          oneOf:
            - type: number
              minimum: 0
              example: 29.99
            - type: object
              properties:
                monthly:
                  type: number
                  minimum: 0
                  example: 29.99
                yearly:
                  type: number
                  minimum: 0
                  example: 299.99
        maxLinks:
          type: integer
          minimum: -1
          example: 500
        maxClicks:
          type: integer
          minimum: -1
          example: 50000
        interval:
          type: string
          enum: [month, year]
          example: month
        isActive:
          type: boolean
          example: true
        isDefault:
          type: boolean
          example: false
        sortOrder:
          type: integer
          example: 2
      required:
        - name
        - price
    UpdatePlanRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
          example: Pro Plan Updated
        slug:
          type: string
          minLength: 2
          maxLength: 50
          pattern: "^[a-z0-9-]+$"
          example: pro-plan
        type:
          type: string
          enum: [free, basic, pro, enterprise]
          example: pro
        description:
          type: string
          example: Updated description
        features:
          $ref: "#/components/schemas/PlanFeatures"
        featuresList:
          type: array
          items:
            $ref: "#/components/schemas/FeatureItem"
        price:
          oneOf:
            - type: number
              minimum: 0
              example: 39.99
            - type: object
              properties:
                monthly:
                  type: number
                  minimum: 0
                  example: 39.99
                yearly:
                  type: number
                  minimum: 0
                  example: 399.99
        maxLinks:
          type: integer
          minimum: -1
          example: 1000
        maxClicks:
          type: integer
          minimum: -1
          example: 100000
        interval:
          type: string
          enum: [month, year]
          example: month
        isActive:
          type: boolean
          example: true
        isDefault:
          type: boolean
          example: false
        sortOrder:
          type: integer
          example: 2
    PlanResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/Plan"
    PlanListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Plan"
        meta:
          $ref: "#/components/schemas/PaginationMeta"

    # Subscription Schemas
...
    CreateSubscriptionRequest:
      type: object
      properties:
        planId:
          type: string
          example: 507f1f77bcf86cd799439011
        billingCycle:
          type: string
          enum: [monthly, yearly, lifetime]
          example: monthly
        paymentProvider:
          type: string
          enum: [stripe, paypal, manual]
          example: stripe
        externalSubscriptionId:
          type: string
          example: sub_1234567890
      required:
        - planId
        - billingCycle
    UpdateSubscriptionRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, cancelled, expired, pending, past_due]
          example: active
        currentPeriodEnd:
          type: string
          format: date-time
          example: "2024-03-15T00:00:00Z"
        cancelAtPeriodEnd:
          type: boolean
          example: true
    SubscriptionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/Subscription"
    SubscriptionListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    SubscriptionHistoryResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"

    # Payment Schemas
...